#refactoring: (refactor) renamed folders with kebab-case and files with camelCase
