+++ packages/web/src/client-descriptors/architecture.ts
