+++ packages/web/src/client-descriptors/canvas/images.ts
