+++ packages/web/src/client-descriptors/canvas/index.ts
