+++ packages/web/src/client-descriptors/canvas/utils.ts
