+++ packages/web/src/client-descriptors/colorDepth.ts
