+++ packages/web/src/client-descriptors/cookiesEnabled.ts
