+++ packages/web/src/client-descriptors/cpuClass.ts
