+++ packages/web/src/client-descriptors/deviceMemory.ts
