+++ packages/web/src/client-descriptors/hardwareConcurrency.ts
