+++ packages/web/src/client-descriptors/index.ts
@@ -1,41 +1,41 @@
 import { getArchitecture } from './architecture';
 import { getAudioFingerprint } from './audio/index';
-import { getAudioContextBaseLatency } from './audio_base_latency';
+import { getAudioContextBaseLatency } from './audioBaseLatency';
 import { getCanvasFingerprint } from './canvas/index';
-import { getColorDepth } from './color_depth';
-import { getColorGamut } from './color_gamut';
-import { getContrastPreference } from './contrast_preference';
-import { areCookiesEnabled } from './cookies_enabled';
-import { getCpuClass } from './cpu_class';
-import { getDeviceMemory } from './device_memory';
-import { getFontPreferences } from './font_preference';
+import { getColorDepth } from './colorDepth';
+import { getColorGamut } from './colorGamut';
+import { getContrastPreference } from './contrastPreference';
+import { areCookiesEnabled } from './cookiesEnabled';
+import { getCpuClass } from './cpuClass';
+import { getDeviceMemory } from './deviceMemory';
+import { getFontPreferences } from './fontPreference';
 import { getFonts } from './fonts';
-import { areColorsForced } from './forced_colors';
-import { getHardwareConcurrency } from './hardware_concurrency';
+import { areColorsForced } from './forcedColors';
+import { getHardwareConcurrency } from './hardwareConcurrency';
 import { isHDR } from './hdr';
-import { getHighEntropyValues } from './hight_entropy_values';
-import { isHaveIndexedDB } from './indexed_db';
-import { areColorsInverted } from './inverted_colors';
+import { getHighEntropyValues } from './hightEntropyValues';
+import { isHaveIndexedDB } from './indexedDB';
+import { areColorsInverted } from './invertedColors';
 import { getLanguages } from './languages';
-import { isHaveLocalStorage } from './local_storage';
+import { isHaveLocalStorage } from './localStorage';
 import { getMathResults } from './math';
-import { getMonochromeDepth } from './monochrome_depth';
-import { isHaveOpenDatabase } from './open_database';
-import { getOsCpu } from './os_cpu';
-import { isPdfViewerEnabled } from './pdf_viewer_enabled';
+import { getMonochromeDepth } from './monochromeDepth';
+import { isHaveOpenDatabase } from './openDatabase';
+import { getOsCpu } from './osCPU';
+import { isPdfViewerEnabled } from './pdfViewerEnabled';
 import { getPlatform } from './platform';
 import { getPlugins } from './plugins';
-import { isMotionReduced } from './reduced_motion';
-import { isTransparencyReduced } from './reduced_transparency';
-import { getScreenFrame } from './screen_frame';
-import { getScreenResolution } from './screen_resolution';
-import { isHaveSessionStorage } from './session_storage';
+import { isMotionReduced } from './reducedMotion';
+import { isTransparencyReduced } from './reducedTransparency';
+import { getScreenFrame } from './screenFrame';
+import { getScreenResolution } from './screenResolution';
+import { isHaveSessionStorage } from './sessionStorage';
 import { getTimezone } from './timezone';
-import { getTouchSupport } from './touch_support';
-import { getUserAgent } from './user_agent';
-import { getUserAgentData } from './user_agent_data';
+import { getTouchSupport } from './touchSupport';
+import { getUserAgent } from './userAgent';
+import { getUserAgentData } from './userAgentData';
 import { getVendor } from './vendor';
-import { getVendorFlavors } from './vendor_flavors';
+import { getVendorFlavors } from './vendorFlavors';
 import { getWebGL } from './webgl/index';
 import { getWebGPU } from './webgpu';
