+++ packages/web/src/client-descriptors/monochromeDepth.ts
