+++ packages/web/src/client-descriptors/osCPU.ts
