+++ packages/web/src/client-descriptors/pdfViewerEnabled.ts
