+++ packages/web/src/client-descriptors/reducedTransparency.ts
