+++ packages/web/src/client-descriptors/screenResolution.ts
