+++ packages/web/src/client-descriptors/vendor.ts
