+++ packages/web/src/client-descriptors/vendorFlavors.ts
