+++ packages/web/src/client-descriptors/webgl/canvas.ts
