+++ packages/web/src/client-descriptors/webgl/utils.ts
