+++ packages/web/src/client-descriptors/webgpu.ts
