+++ packages/web/src/index.ts
@@ -1,10 +1,7 @@
 import { FPClient } from './FPClient';
 
-export type { Descriptors } from './clientDescriptors/index';
-export type {
-  DescriptorsRes,
-  FPClientInterface,
-} from './types/index';
+export type { Descriptors } from './client-descriptors/index';
+export type { DescriptorsRes, FPClientInterface } from './types/index';
 
 export { FPClient };
 export default FPClient;
